Slight restructuring of the code within lenstronomy
